requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2
PyJWT
pytz
sqlparse
tzdata
Pillow
python-docx